requirements.txt
Pillow
pytest
slackclient
upsidedown>=0.4
